Unify tese cleaning and document LLM configuration
