+++ .env.example
@@ -1,5 +1,5 @@
 LLM_PROVIDER=openai
 LLM_MODEL=gpt-4.1-mini
 LLM_API_KEY=sk-your-key
-# Opcional: ajuste URL base se usar proxy ou Azure.
+# Opcional para proxies/self-host: ajuste a URL base.
 # LLM_BASE_URL=https://api.openai.com/v1/chat/completions
