+++ .gitignore
@@ -159,6 +159,11 @@ venv.bak/
 .dmypy.json
 dmypy.json
 
+# Project artifacts
+runs/
+data/raw/
+reports/*.csv
+
 # Pyre type checker
 .pyre/
